+++ backend/routes/api_user_ItemCategory.js
@@ -32,6 +32,7 @@ const getUserItemCategoryById = async (req, res) => {
         res.send(itemCategory ? itemCategory : {"message":"No record found"});
     } catch (err) {
         console.log(err);
+        res.sendStatus(400);
     }
 }
  
@@ -56,11 +57,25 @@ const createUserItemCategory = async (req, res) => {
         });
     } catch (err) {
         console.log(err);
+        res.sendStatus(400);
     }
 }
  
 // Update ItemCategory by id
 const updateUserItemCategory = async (req, res) => {
+
+    if(!req.auth.user.id){res.sendStatus(401);return false;}
+    if(!req.params.id){res.sendStatus(401);return false;}
+    if(req.body.id && (req.body.id != req.params.id)){res.sendStatus(401);return false;}
+    // console.log(req.body);
+    if(!req.body.department_id){res.send({"message":"department_id is needed in params"});return false;}
+    // var itemcheck = await Item.findByPk(req.body.id);
+    // console.log(req.body);
+    var departmentId = req.body.department_id;
+    if(!departmentId){res.send({"message":"departmentId of record not found"});return false;}
+    console.log(departmentId);
+    if(!await userHasAccessToDepartment(req.auth,departmentId)){res.sendStatus(403);return false;}
+
     try {
         await ItemCategory.update(req.body, {
             where: {
@@ -72,11 +87,21 @@ const updateUserItemCategory = async (req, res) => {
         });
     } catch (err) {
         console.log(err);
+        res.sendStatus(400);
     }
 }
  
 // Delete ItemCategory by id
 const deleteUserItemCategory = async (req, res) => {
+
+    if(!req.auth.user.id){res.sendStatus(401);return false;}
+    if(!req.params.id){res.send({"message":"Id is needed in params"});return false;}
+    var itemcheck = await ItemCategory.findByPk(req.params.id);
+    var departmentId = itemcheck.department_id;
+    if(!departmentId){res.send({"message":"departmentId of record not found"});return false;}
+    console.log(departmentId);
+    if(!await userHasAccessToDepartment(req.auth,departmentId)){res.sendStatus(403);return false;}
+
     try {
         await ItemCategory.destroy({
             where: {
@@ -88,6 +113,7 @@ const deleteUserItemCategory = async (req, res) => {
         });
     } catch (err) {
         console.log(err);
+        res.sendStatus(400);
     }
 }
