+++ backend/routes/api_user_Room.js
@@ -32,6 +32,7 @@ const getUserRoomById = async (req, res) => {
         res.send(room ? room : {"message":"No record found"});
     } catch (err) {
         console.log(err);
+        res.sendStatus(400);
     }
 }
  
@@ -56,11 +57,25 @@ const createUserRoom = async (req, res) => {
         });
     } catch (err) {
         console.log(err);
+        res.sendStatus(400);
     }
 }
  
 // Update Room by id
 const updateUserRoom = async (req, res) => {
+
+    if(!req.auth.user.id){res.sendStatus(401);return false;}
+    if(!req.params.id){res.sendStatus(401);return false;}
+    if(req.body.id && (req.body.id != req.params.id)){res.sendStatus(401);return false;}
+    // console.log(req.body);
+    if(!req.body.department_id){res.send({"message":"department_id is needed in params"});return false;}
+    // var itemcheck = await Item.findByPk(req.body.id);
+    // console.log(req.body);
+    var departmentId = req.body.department_id;
+    if(!departmentId){res.send({"message":"departmentId of record not found"});return false;}
+    console.log(departmentId);
+    if(!await userHasAccessToDepartment(req.auth,departmentId)){res.sendStatus(403);return false;}
+
     try {
         await Room.update(req.body, {
             where: {
@@ -72,11 +87,21 @@ const updateUserRoom = async (req, res) => {
         });
     } catch (err) {
         console.log(err);
+        res.sendStatus(400);
     }
 }
  
 // Delete Room by id
 const deleteUserRoom = async (req, res) => {
+
+    if(!req.auth.user.id){res.sendStatus(401);return false;}
+    if(!req.params.id){res.send({"message":"Id is needed in params"});return false;}
+    var roomcheck = await Room.findByPk(req.params.id);
+    var departmentId = roomcheck.department_id;
+    if(!departmentId){res.send({"message":"departmentId of record not found"});return false;}
+    console.log(departmentId);
+    if(!await userHasAccessToDepartment(req.auth,departmentId)){res.sendStatus(403);return false;}
+
     try {
         await Room.destroy({
             where: {
@@ -88,6 +113,7 @@ const deleteUserRoom = async (req, res) => {
         });
     } catch (err) {
         console.log(err);
+        res.sendStatus(400);
     }
 }
